fetch latest

+++ app/infra/Cache.ts
@@ -4,17 +4,19 @@ import { Connection } from "./Connection";
 
 console.log(Conf.platform);
 
-export const read = async (path) => {
+export const read = async (path, forceSync:boolean = true) => {
     if(!Connection.isOnline){
         console.log('User offline, reading from cache');
     }
-    if(Connection.isOnline){
+    const fromCache = await AsyncStorage.getItem(path);
+
+    if(Connection.isOnline && !(!forceSync && fromCache)){
         const response = await fetch(`${Conf.platform}${path}`);
         const data = await response.json();
         await AsyncStorage.setItem(path, JSON.stringify(data));
         return data;
     }
-    const fromCache = await AsyncStorage.getItem(path);
+    
     if(fromCache){
         return JSON.parse(fromCache);
     }
